requirements.txt
memory-profiler==0.61.0
numpy==1.25.2
Pillow==10.0.0
psutil==5.9.5
